+++ src/main/java/lotto/entity/PrizePolicy.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.entity;
 
 import java.util.Arrays;
