+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.view;
 
-import java.util.List;
 import java.util.Scanner;
 
 public class InputView {
